+++ usage/update/github.go
@@ -24,6 +24,10 @@ type githubRelease struct {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+var githubAPI = "https://api.github.com"
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // GitHubChecker checks new releases on GitHub
 func GitHubChecker(app, version, data string) (string, time.Time, bool) {
 	if version == "" || data == "" {
@@ -45,12 +49,12 @@ func GitHubChecker(app, version, data string) (string, time.Time, bool) {
 func getLatestGitHubRelease(app, version, repository string) *githubRelease {
 	engine := req.Engine{}
 
-	engine.SetDialTimeout(1)
-	engine.SetRequestTimeout(1)
+	engine.SetDialTimeout(3)
+	engine.SetRequestTimeout(3)
 	engine.SetUserAgent(app, version, "go.ek")
 
 	response, err := engine.Get(req.Request{
-		URL:         "https://api.github.com/repos/" + repository + "/releases/latest",
+		URL:         githubAPI + "/repos/" + repository + "/releases/latest",
 		AutoDiscard: true,
 	})